24-06-2025/backend/requirements.txt
langchain>=0.1.0
google-generativeai>=0.3.0
chromadb>=0.4.22
beautifulsoup4>=4.12.0
lxml>=4.9.0
cssselect>=1.2.0
tiktoken>=0.5.2
pydantic>=2.5.0
langserve>=0.0.30
langgraph>=0.0.20
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
cssutils>=2.9.0
esprima>=4.0.1
langchain-community
langchain_google_genai
pymupdf

